move utilities file to utils folder
